+++ include/cltune/kernel_info.h
@@ -38,7 +38,7 @@
 #include <functional>
 
 #include "cl.hpp"
-#include "internal/string_range.h"
+#include "cltune/string_range.h"
 
 namespace cltune {
 // =================================================================================================
