+++ include/cltune/searchers/full_search.h
@@ -5,7 +5,8 @@
 //
 // Author: cedric.nugteren@surfsara.nl (Cedric Nugteren)
 //
-// This file implements the simulated annealing algorithm. (...)
+// This file implements a full-search algorithm, testing all configurations exhaustively. It is
+// derived from the basic search class Searcher.
 //
 // -------------------------------------------------------------------------------------------------
 //
@@ -25,28 +26,20 @@
 //
 // =================================================================================================
 
-#ifndef CLTUNE_SEARCHERS_ANNEALING_H_
-#define CLTUNE_SEARCHERS_ANNEALING_H_
+#ifndef CLTUNE_SEARCHERS_FULL_SEARCH_H_
+#define CLTUNE_SEARCHERS_FULL_SEARCH_H_
 
 #include <vector>
-#include <random>
 
-#include "internal/searcher.h"
+#include "cltune/searcher.h"
 
 namespace cltune {
 // =================================================================================================
 
 // See comment at top of file for a description of the class
-class Annealing: public Searcher {
+class FullSearch: public Searcher {
  public:
-
-  // Maximum number of successive visits to already visited states. If this number is exceeded, the
-  // algorithm ends
-  static constexpr auto kMaxAlreadyVisitedStates = 10;
-
-  // Takes additionally a fraction of configurations to consider
-  Annealing(const Configurations &configurations,
-            const double fraction, const double max_temperature);
+  FullSearch(const Configurations &configurations);
 
   // Retrieves the next configuration to test
   virtual KernelInfo::Configuration GetConfiguration() override;
@@ -57,39 +50,11 @@ class Annealing: public Searcher {
   // Retrieves the total number of configurations to try
   virtual size_t NumConfigurations() override;
 
-  // Pushes feedback (in the form of execution time) from the tuner to the search algorithm
-  virtual void PushExecutionTime(const double execution_time) override;
-
  private:
-
-  // Retrieves a vector with all neighbours of a reference configuration
-  std::vector<size_t> GetNeighboursOf(const size_t reference_id) const;
-
-  // Computes the acceptance probability P of simulated annealing based on the 'energy' of the
-  // current and neighbouring state, and the 'temperature'.
-  double AcceptanceProbability(const double current_energy,
-                               const double neighbour_energy,
-                               const double temperature) const;
-
-  // Configuration parameters
-  double fraction_;
-  double max_temperature_;
-
-  // Annealing-specific member variables
-  size_t num_visited_states_;
-  size_t current_state_;
-  size_t neighbour_state_;
-  size_t num_already_visisted_states_;
-
-  // Random number generation
-  std::random_device rd_;
-  std::default_random_engine generator_;
-  std::uniform_int_distribution<int> int_distribution_;
-  std::uniform_real_distribution<double> probability_distribution_;
 };
 
 // =================================================================================================
 } // namespace cltune
 
-// CLTUNE_SEARCHERS_ANNEALING_H_
+// CLTUNE_SEARCHERS_FULL_SEARCH_H_
 #endif
