+++ src/memory.cc
@@ -25,7 +25,7 @@
 //
 // =================================================================================================
 
-#include "internal/memory.h"
+#include "cltune/memory.h"
 
 namespace cltune {
 // =================================================================================================
