+++ src/opencl.cc
@@ -25,7 +25,7 @@
 //
 // =================================================================================================
 
-#include "internal/opencl.h"
+#include "cltune/opencl.h"
 
 namespace cltune {
 // =================================================================================================
