+++ src/searchers/full_search.cc
@@ -25,7 +25,7 @@
 //
 // =================================================================================================
 
-#include "internal/searchers/full_search.h"
+#include "cltune/searchers/full_search.h"
 
 namespace cltune {
 // =================================================================================================
