+++ src/string_range.cc
@@ -25,7 +25,7 @@
 //
 // =================================================================================================
 
-#include "internal/string_range.h"
+#include "cltune/string_range.h"
 
 #include <string>
