+++ test/kernel_info.cc
@@ -25,7 +25,7 @@
 //
 // =================================================================================================
 
-#include "internal/kernel_info.h"
+#include "cltune/kernel_info.h"
 
 #include <memory>
